% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{choose_first_installed_font}
\alias{choose_first_installed_font}
\title{Choose the first installed out of the list of choices}
\usage{
choose_first_installed_font(choices, verbose = FALSE)
}
\arguments{
\item{choices}{character, vector of font family names}

\item{verbose}{logical, whether to display chosen font}
}
\value{
the first installed font from \code{choices}, or
failing that, "sans"
}
\description{
As a side-effect, a warning will be thrown if none of the
choices are installed.
}
\examples{
if (interactive()) {
  print(choose_first_installed_font(c("Blah blah", "Arial", "Another made-up font")))
}
}
