% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{color_text}
\alias{color_text}
\title{Color a specific piece of text with a specific RGB hexcode.}
\usage{
color_text(color, text)
}
\arguments{
\item{text}{}
}
\value{

}
\description{
This is often useful within a \code{glue::glue} function for
a graph title or subtitle.
}
\examples{
color_text("#190C65", "predicted")
}
