% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{demo_theme_ku}
\alias{demo_theme_ku}
\title{Demonstrate \link{theme_ku} compared to other themes with a
simple scatter plot (or other of your choice)}
\usage{
demo_theme_ku(plot = NULL)
}
\arguments{
\item{plot}{\code{\link[ggplot2]{ggplot2}} plot, plot to
use for comparision. (Default: \code{NULL} indicates that
the default scatter plot should be used)}
}
\value{
\code{\link[ggplot2]{ggplot2}} plot comparing
four different themes
}
\description{
Demonstrate \link{theme_ku} compared to other themes with a
simple scatter plot (or other of your choice)
}
\examples{
if (interactive()) {
  demo_theme_ku()
}
}
