% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{ku_color}
\alias{ku_color}
\title{Extract colors from the KU palette}
\usage{
ku_color(...)
}
\arguments{
\item{...}{character names of colors. See \code{ku_color()}
for available colors.}
}
\value{
named character vector of colors
}
\description{
Extract colors from the KU palette
}
\examples{
if (interactive()) {
  print(ku_color())
  print(ku_color("KU Blue", "Night"))
}
}
\seealso{
show_colors_with_names can display the palette.
}
