% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{ku_color_text}
\alias{ku_color_text}
\title{Color a specific piece of text with a KU color}
\usage{
ku_color_text(color_name, text, ...)
}
\arguments{
\item{color_name}{character, valid KU color}

\item{text}{character, text to color}

\item{...}{passed to \link{color_text}}
}
\value{
character, HTML-span-wrapped version of text
styled according to KU color \code{color_name} (and
\code{weight})
}
\description{
This is often useful within a \link[glue:glue]{glue::glue} function for a
graph title or subtitle.
}
\examples{
ku_color_text("Crimson", "predicted")
}
