% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{ku_get_version}
\alias{ku_get_version}
\title{Get version of KU brand {kubrand} package}
\usage{
ku_get_version()
}
\value{
character version string
}
\description{
Get version of KU brand {kubrand} package
}
\examples{
ku_get_version()
if (interactive()) {
  ku_get_version()
}
}
