% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ku_pal}
\alias{ku_pal}
\title{Access/interpolate from one of the menu of KU palettes}
\usage{
ku_pal(palette = "full", reverse = FALSE, ...)
}
\arguments{
\item{palette}{character vector. See \code{ku_palettes}.}

\item{reverse}{boolean. Indicates whether to reverse the
order.}

\item{...}{Additional arguments to pass to \code{colorRampPalette()}}
}
\value{
a function taking a number of colors and
returning a palette
}
\description{
Access/interpolate from one of the menu of KU palettes
}
\examples{
if (interactive()) {
  my_palette <- ku_pal("div_red_blue")
  my_palette(8)
  show_colors_with_names(my_palette(7), ncol = 1)
}
}
