% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubrand-package.R
\docType{package}
\name{kubrand-package}
\alias{kubrand}
\alias{kubrand-package}
\title{kubrand: KU Branding for ggplot2}
\description{
Do you want to make a plot with ggplot2 but with KU colors?
    Then install this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bvancilku/kubrand}
  \item Report bugs at \url{https://github.com/bvancilku/kubrand/issues}
}

}
\author{
\strong{Maintainer}: Brian Vancil \email{43145327+bvancilku@users.noreply.github.com} (\href{https://orcid.org/0000-0003-1145-3681}{ORCID})

Authors:
\itemize{
  \item Nathan Kuhn \email{blanketyblank@users.noreply.github.com}
}

}
\keyword{##}
\keyword{-}
\keyword{Useful}
\keyword{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
\keyword{https://github.com/hneth/unikn}
\keyword{internal}
\keyword{resources:}
