% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label_year_range}
\alias{label_year_range}
\title{Label special year type (like an academic or fiscal
year), which might span calendar years}
\usage{
label_year_range(prefix = "", suffix = "", digits = NULL, ...)
}
\arguments{
\item{prefix}{character, prefix to use for all labels
(default: "")}

\item{suffix}{character, suffix to use for all labels
(default: "")}

\item{digits}{integer, number of digits to use for second
number in range (default: NULL, which indicates to
guess)}

\item{...}{not used}
}
\value{
function that takes an integer and returns a
string
}
\description{
\itemize{
\item Special years are assumed to be "named" by the end
year. Simply add one to your input if this is not the
case. - This is intended to be used with
\code{\link[ggplot2]{ggplot2}} like other labellers, e.g.
\link[scales:label_percent]{scales::label_percent}.
}
}
\examples{
if (interactive()) {
  label_year_range(prefix = "FY ", digits = 4L)(2001:2010)
  label_year_range(prefix = "FY ")(2001:2010)
}
}
