% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plus-replace.R
\name{\%+replace\%}
\alias{\%+replace\%}
\title{Modify the active theme (imported)}
\usage{
e1 \%+replace\% e2
}
\arguments{
\item{e1}{original ggplot2 theme}

\item{e2}{ggplot2 theme element updates}
}
\value{
Updated
}
\description{
See \code{ggplot2::\link[ggplot2:theme_get]{\%+replace\%}} for details.
}
\keyword{internal}
