% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plot_with_fonts}
\alias{save_plot_with_fonts}
\title{Alternative to \link[ggplot2:ggsave]{ggplot2::ggsave} that embeds fonts}
\usage{
save_plot_with_fonts(plot, path, ...)
}
\arguments{
\item{plot}{\code{\link[ggplot2]{ggplot2}} plot to save}

\item{path}{character, path (relative to current directory)}

\item{...}{other arguments passed to \link[ggplot2:ggsave]{ggplot2::ggsave}.}
}
\value{
\code{path}
}
\description{
Alternative to \link[ggplot2:ggsave]{ggplot2::ggsave} that embeds fonts
}
\examples{
if (interactive()) {
  save_plot_with_fonts(demo_theme_ku(), "test.pdf")
}
}
