% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{show_colors_with_names}
\alias{show_colors_with_names}
\title{Show a vector of hexcode colors using names.}
\usage{
show_colors_with_names(
  colours,
  labels = TRUE,
  borders = NULL,
  cex_label = 1,
  ncol = NULL
)
}
\arguments{
\item{colours}{character, vector of RGB hex codes with names for the colors}

\item{labels}{logical, whether to display color name labels}

\item{borders}{color to show for borders}

\item{cex_label}{size for label}

\item{ncol}{number of columns (default: \code{NULL} means to make the output as square as possible.)}
}
\description{
Modelled off of \code{scales::show_col} For an alternative to
\code{show_colors_with_names(ku_palette)}, try
\code{colorspace::swatchplot(as.list(ku_palette))}
}
\examples{
if (interactive()) {
  show_colors_with_names(ku_palette)
}
}
