% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{theme_ku}
\alias{theme_ku}
\title{KU branded {ggplot2} theme}
\usage{
theme_ku(
  base_size = 10,
  base_family = "sans",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  title_location = "plot"
)
}
\arguments{
\item{title_location}{character, c("plot", "pane")}
}
\value{

}
\description{
TODO:
\itemize{
\item Think about effect of \code{ggplot2::coord_cartesian(expand = FALSE)}
\itemize{
\item And also maybe add clip = 'off'
}
\item Add spacing between title and subtitle.
\item Consider adding KU logo watermark in corner (with option in function to turn it off).
\item Lighten the axis tick labels.
}
}
