% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{ku_color_text}
\alias{ku_color_text}
\title{Color a specific piece of text with a KU color}
\usage{
ku_color_text(color_name, text)
}
\arguments{
\item{text}{}
}
\value{

}
\description{
This is often useful within a \code{glue::glue} function for
a graph title or subtitle.
}
\examples{
ku_color_text("Crimson", "predicted")
}
