% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plus-replace.R
\name{\%+replace\%}
\alias{\%+replace\%}
\title{Modify the active theme (imported)}
\usage{
theme1 \%+replace\% element2
}
\arguments{
\item{theme1}{original ggplot2 theme}

\item{element2}{ggplot2 theme element updates}
}
\value{
Updated
}
\description{
See \code{ggplot2::\link[ggplot2:theme_get]{\%+replace\%}} for details.
}
\keyword{internal}
